Move UserRole file to new folder
