+++ src/main/java/com/k9club/api/controller/AuthController.java
@@ -3,9 +3,9 @@ package com.k9club.api.controller;
 import com.k9club.api.dao.UserDao;
 import com.k9club.api.dto.LoginRequestDTO;
 import com.k9club.api.model.User;
+import com.k9club.api.model.enums.UserRole;
 import com.k9club.api.security.AppUserDetails;
 import com.k9club.api.security.ISecurityUtils;
-import com.k9club.api.security.UserRole;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
